Modifications & changes to the UI
Changed the structure of the files & improvements to the UI
